+++ libs/RadioHead/RHutil/simulator.h
@@ -37,16 +37,16 @@ public:
     // TODO: move these from being inlined
     void begin(int baud) {}
 
-    size_t println(const char* s)
+    int32_t println(const char* s)
     {
 	print(s);
 	return printf("\n");
     }
-    size_t print(const char* s)
+    int32_t print(const char* s)
     {
 	return printf("%s", s); // This style prevent warnings from [-Wformat-security]
     }
-    size_t print(unsigned int n, int base = DEC)
+    int32_t print(unsigned int n, int base = DEC)
     {
 	if (base == DEC)
 	    return printf("%d", n);
@@ -58,19 +58,19 @@ public:
 	else
 	    return 0;
     }
-    size_t print(char ch)
+    int32_t print(char ch)
     {
         return printf("%c", ch);
     }
-    size_t println(char ch)
+    int32_t println(char ch)
     {
         return printf("%c\n", ch);
     }
-    size_t print(unsigned char ch, int base = DEC)
+    int32_t print(unsigned char ch, int base = DEC)
     {
 	return print((unsigned int)ch, base);
     }
-    size_t println(unsigned char ch, int base = DEC)
+    int32_t println(unsigned char ch, int base = DEC)
     {
 	print((unsigned int)ch, base);
 	return printf("\n");
