+++ libs/RadioHead/STM32ArduinoCompat/HardwareSerial.cpp
@@ -240,7 +240,7 @@ int HardwareSerial::read(void)
     return _rxRingBuffer.read();
 }
 
-size_t HardwareSerial::write(uint8_t ch)
+int32_t HardwareSerial::write(uint8_t ch)
 {
     _txRingBuffer.write(ch); // Queue it
     USART_ITConfig(_usart, USART_IT_TXE, ENABLE); // Enable the TX interrupt
