+++ src/RHutil/RasPi.h
@@ -49,13 +49,13 @@ class SerialSimulator
 
     // TODO: move these from being inlined
     static void begin(int baud);
-    static size_t println(const char *s);
-    static size_t print(const char *s);
-    static size_t print(unsigned int n, int base = DEC);
-    static size_t print(char ch);
-    static size_t println(char ch);
-    static size_t print(unsigned char ch, int base = DEC);
-    static size_t println(unsigned char ch, int base = DEC);
+    static int32_t println(const char *s);
+    static int32_t print(const char *s);
+    static int32_t print(unsigned int n, int base = DEC);
+    static int32_t print(char ch);
+    static int32_t println(char ch);
+    static int32_t print(unsigned char ch, int base = DEC);
+    static int32_t println(unsigned char ch, int base = DEC);
 };
 
 extern SerialSimulator Serial;
